+++ Manager/ModLoaderBridge.cs
@@ -403,12 +403,6 @@ namespace Sheepy.Modnix.MainGUI {
                txt.Text = "\rRuntime error(s) detected on last run, may be not safe to use."; break;
             case "runtime_warning" :
                txt.Text = "\rRuntime warning(s) detected on last run."; break;
-            case "config_mismatch" :
-               txt.Text = "\rDefaultConfig different from new instance defaults."; break;
-         case "unspported_flags" :
-            txt.Text = "Mod Flags requires Modnix 3 or above."; break;
-         case "unspported_actions" :
-            txt.Text = "Mod Actions requires Modnix 3 or above.\rMod may not work or only partially work."; break;
             default :
                txt.Text = "\r" + notice.Message.ToString(); break;
          }