+++ Manager/Resources/Changelog.md
@@ -4,7 +4,8 @@ Modnix Changelog
 
 * New: Show PPML version in app status.
 * New: Mod Actions - a new system that allows mods to declare Actions and/or to support them.
-* Mod: Removed deprecated logger and config param from mod initialisers. Use API instead.
+* Mod: Logger and config param no longer supported for mod initialisers and will be passed null.  Use API instead.
+* Mod: DefaultConfig is no longer supported.  Use ConfigType instead.
 * Log: Mod loading messages are now logged with the mod's logger.
 * Other: Mod phases other than SplashMod now depends on correct Harmony to function. SplashMods can still be loaded without Harmony.